requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
openai
faiss-cpu
pypdf